refactor: move code generation logic to generator.service
